frontend: isolate the fronted JS package here
